ai-healthcare-assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-openai==0.0.5
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pypdf==3.17.4
unstructured[local-inference]==0.11.2
openai==1.3.7
pyyaml==6.0.1